requirements.txt
aiohttp
aiofiles
beautifulsoup4
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv